Added uninstall hooks for a python package

+++ sbi.tcl
@@ -70,13 +70,13 @@ for {set i 0} {$i < $argc} {incr i} {
 			incr i [llength $build_reps]
 			incr i -1
 		}
-        --sbi-dir {
-            # Changes the folder where sbi stores installed packages and recipes
+		--sbi-dir {
+			# Changes the folder where sbi stores installed packages and recipes
 			incr i
 			set sbi_dir [lindex $argv $i]
-            set sbi_dir [file normalize $sbi_dir]
-            puts "Setting the SBI storage folder to $sbi_dir"
-        }
+			set sbi_dir [file normalize $sbi_dir]
+			puts "Setting the SBI storage folder to $sbi_dir"
+		}
 		--rebuild-deps {
 			set rebuild_deps 1
 		}
@@ -89,9 +89,9 @@ for {set i 0} {$i < $argc} {incr i} {
 			incr i [llength $import_reps]
 			incr i -1
 		}
-        --rm-old-builds {
+		--rm-old-builds {
 			set rm_old_builds 1
-        }
+		}
 		--delete {
 			incr i
 			set del_reps [get_arg_list $argv $i]
@@ -118,7 +118,7 @@ file mkdir $rep_dir
 
 proc get_pkg_dir {pkg_name_ver} {
 	global inst_dir
-    return [file join $inst_dir $pkg_name_ver]
+	return [file join $inst_dir $pkg_name_ver]
 }
 
 proc exec_stdout {exec_str} {
@@ -156,19 +156,36 @@ if {[llength $del_reps] > 0} {
     }
 }
 
+proc get_rep_short_name {rep_dict} {
+	set short_name "[dict get $rep_dict plat]-[dict get $rep_dict name]-[dict get $rep_dict ver]"
+	return $short_name
+}
+
+proc get_rep_path {rep_short_name} {
+	global rep_dir
+	return [file join $rep_dir $rep_short_name.tcl]
+}
+
 proc import_rep_file {f_import} {
 	source $f_import
-	set short_name "[dict get $rep_info plat]-[dict get $rep_info name]-[dict get $rep_info ver]"
+	set short_name [get_rep_short_name $rep_info]
 	global rep_dir
-	set rep_dest [file join $rep_dir $short_name.tcl]
+	set rep_dest [get_rep_path $short_name]
 	file copy -force $f_import $rep_dest
 }
 
 if {[llength $import_reps] > 0} {
 	foreach rep $import_reps {
 		import_rep_file $rep
-        puts "Imported $rep"
+		puts "Imported $rep"
 	}
+}
+
+proc proc_exists {file name} {
+    set search_f [open $file r]
+    set f_text [read $search_f]
+    close $search_f
+    return [regexp "proc\\s+$name\\s+\{" $f_text]
 }
 
 if {[llength $rem_reps] > 0} {
@@ -181,8 +198,29 @@ if {[llength $rem_reps] > 0} {
 				file delete -force -- $folder
 			}
 		} else {
+			set rep_file [get_rep_path $r_rep]
+			set run_uninstall 0
+			if {[file isfile $rep_file]} {
+				source $rep_file
+				set run_uninstall [proc_exists $rep_file uninstall]
+			}
 			set exp_dir [file join $inst_dir $r_rep]
 			if {[file isdirectory $exp_dir]} {
+				if {$run_uninstall} {
+					set old_path $::env(PATH)
+					set build_needs {}
+					if {[dict exists $rep_info rm_needs]} {
+						foreach need [dict get $rep_info rm_needs] {
+							set need_bin_dir [file join $inst_dir $need bin]
+							if {[file isdirectory $need_bin_dir]} {
+								puts "Adding $need_bin_dir to PATH for uninstall"
+								set env(PATH) "$need_bin_dir:$::env(PATH)"
+							}
+						}
+					}
+					uninstall $r_rep $exp_dir
+					set env(PATH) $old_path
+				}
 				puts "Removing $r_rep"
 				file delete -force -- $exp_dir
 			} else {
@@ -230,12 +268,6 @@ proc exec_log_cmd {cmd log_path} {
 	close $cmd_log
 }
 
-proc proc_exists {file name} {
-    set search_f [open $file r]
-    set f_text [read $search_f]
-    close $search_f
-    return [regexp "proc\\s+$name\\s+\{" $f_text]
-}
 
 # I just made up this hash function. It might not be any good.
 proc str_to_hash {str} {
@@ -262,15 +294,15 @@ proc build_recipe {rep_path {rebuild 0} {rebuild_deps 0} {do_check 0}} {
 	} else {
 		set src_path $rep_path
 	}
-    if {[file isfile $src_path] == 0} {
-        error "$src_path Isn't a file (or doesn't exist)!. You may need to import the recipe."
-    }
-    source $src_path
-    # Get the hash for the recipe
-    set src_f [open $src_path r]
-    set src_text [read $src_f]
-    set src_hash [str_to_hash $src_text]
-    close $src_f
+	if {[file isfile $src_path] == 0} {
+		error "$src_path Isn't a file (or doesn't exist)!. You may need to import the recipe."
+	}
+	source $src_path
+	# Get the hash for the recipe
+	set src_f [open $src_path r]
+	set src_text [read $src_f]
+	set src_hash [str_to_hash $src_text]
+	close $src_f
 
 	# Assume data is in rep_info 
 	puts "Recipie info:"
@@ -283,22 +315,25 @@ proc build_recipe {rep_path {rebuild 0} {rebuild_deps 0} {do_check 0}} {
 	set exp_path [file join $inst_dir $short_name]
 	if {[file isdirectory $exp_path] && $rebuild == 0} {
 		puts "$short_name is installed, skipping build"
-        set ret [list $short_name]
+		set ret [list $short_name]
 		return $ret
 	}
-	set srcs [dict get $rep_info srcs]
-	global src_dir
-	foreach src $srcs {
-		set src_name [file tail $src]
-		set save_path [file join $src_dir $src_name]
-		if {[file exists $save_path] == 0} {
-			puts "Downloading $src_name ($src) to $save_path"
-			exec >&@stdout wget -O $save_path $src
-			puts "Downloaded $src_name to $save_path"
+
+	if {[dict exists $rep_info srcs]} {
+		set srcs [dict get $rep_info srcs]
+		global src_dir
+		foreach src $srcs {
+			set src_name [file tail $src]
+			set save_path [file join $src_dir $src_name]
+			if {[file exists $save_path] == 0} {
+				puts "Downloading $src_name ($src) to $save_path"
+				exec >&@stdout wget -O $save_path $src
+				puts "Downloaded $src_name to $save_path"
+			}
 		}
 	}
 
-    set built_pkgs [list]
+	set built_pkgs [list]
 	# Build needed libraries
 	if {[dict exists $rep_info build_needs]} {
 		set b_needs [dict get $rep_info build_needs]
@@ -306,96 +341,99 @@ proc build_recipe {rep_path {rebuild 0} {rebuild_deps 0} {do_check 0}} {
 			set need_inst_dir [file join $inst_dir $need]
 			if {[file isdirectory $need_inst_dir] == 0} {
 				set pkgs [build_recipe $need 0 0 $do_check]
-            # Make sure we haven't built the package already if we're rebuilding deps
+				# Make sure we haven't built the package already if we're rebuilding deps
 			} elseif {$rebuild_deps && [lsearch $built_pkgs $need] == -1} {
 				set pkgs [build_recipe $need 1 1 $do_check]
 			} else {
-                set pkgs [list $need]
-            }
-            lappend built_pkgs $pkgs
-        }
+				set pkgs [list $need]
+			}
+			lappend built_pkgs $pkgs
+		}
 	}
+	puts "Built/have $built_pkgs for $short_name"
 
 	global build_dir
 	set tmp_build_dir [file join $build_dir $short_name]
-    # Clear out the builds completely before extracting
-    # This keeps the build folder somewhat empty
-    # Don't to this because we can't run concurrent builds
+	# Clear out the builds completely before extracting
+	# This keeps the build folder somewhat empty
+	# Don't to this because we can't run concurrent builds
 	file delete -force -- $tmp_build_dir
 	file mkdir $tmp_build_dir
 
-	# Excract source to build dir
-	foreach src $srcs {
-		set src_name [file tail $src]
-		set save_path [file join $src_dir $src_name]
-		puts "Extracting $save_path to $tmp_build_dir"
-        set cmd "tar -xf $save_path -C $tmp_build_dir"
-        if {[string match "*.zip" $save_path]} {
-            set cmd "unzip $save_path -d $tmp_build_dir"
-        }
-		exec >&@stdout {*}$cmd
+	if {[dict exists $rep_info srcs]} {
+		# Excract source to build dir
+		foreach src $srcs {
+			set src_name [file tail $src]
+			set save_path [file join $src_dir $src_name]
+			puts "Extracting $save_path to $tmp_build_dir"
+			set cmd "tar -xf $save_path -C $tmp_build_dir"
+			if {[string match "*.zip" $save_path]} {
+				set cmd "unzip $save_path -d $tmp_build_dir"
+			}
+			exec >&@stdout {*}$cmd
+		}
 	}
 
-    # Make sure the functions are up to date.
-    # sourcing other file could override them
-    source $src_path
+	# Make sure the functions are up to date.
+	# sourcing other file could override them
+	source $src_path
 
-    set run_prep [proc_exists $src_path prepare]
-    set run_build [proc_exists $src_path build]
-    set run_install [proc_exists $src_path install]
-    set run_check [proc_exists $src_path check]
+	set run_prep [proc_exists $src_path prepare]
+	set run_build [proc_exists $src_path build]
+	set run_install [proc_exists $src_path install]
+	set run_check [proc_exists $src_path check]
 
-    set pkg_inst_dir [file join $inst_dir $short_name]
+	set pkg_inst_dir [file join $inst_dir $short_name]
 
-    set old_env [array get ::env]
-    # Add the build needs to the PATH
-    puts "Built these pkgs $built_pkgs"
-    foreach built_pkg $built_pkgs {
-        set pkg_bin_dir [file join [get_pkg_dir $built_pkg] bin]
-        if {[file isdirectory $pkg_bin_dir]} {
-            puts "Adding $pkg_bin_dir to build PATH"
-            set ::env(PATH) "$pkg_bin_dir:$::env(PATH)"
-        } else {
-            puts "$pkg_bin_dir does not exist"
-        }
-    }
+	set old_env [array get ::env]
+	# Add the build needs to the PATH
+	puts "Built these pkgs $built_pkgs"
+	foreach built_pkg $built_pkgs {
+		set pkg_bin_dir [file join [get_pkg_dir $built_pkg] bin]
+		if {[file isdirectory $pkg_bin_dir]} {
+			puts "Adding $pkg_bin_dir to build PATH"
+			set ::env(PATH) "$pkg_bin_dir:$::env(PATH)"
+		} else {
+			puts "$pkg_bin_dir does not exist"
+		}
+	}
 
-    cd $tmp_build_dir
-    if {$run_prep} {
-        puts "Running prepare{}"
-        prepare $pkg_name $pkg_inst_dir $tmp_build_dir
-    }
-    if {$run_build} {
-        puts "Running build{}"
-        build [dict get $rep_info name] [dict get $rep_info ver] $pkg_inst_dir $tmp_build_dir
-    }
-    if {$run_check && $do_check} {
-        puts "Running check{}"
-        check $short_name $pkg_inst_dir $tmp_build_dir
-    }
-    if {$run_install} {
-        puts "Running install{}"
-        install $short_name $pkg_inst_dir $tmp_build_dir
-    }
-    set inst_files [glob -nocomplain -directory $inst_dir *]]
-    if {[llength $inst_files] == 0} {
-        error "This recipe didn't install any files!"
-    }
+	cd $tmp_build_dir
+	if {$run_prep} {
+		puts "Running prepare{}"
+		prepare $pkg_name $pkg_inst_dir $tmp_build_dir
+	}
+	if {$run_build} {
+		puts "Running build{}"
+		build [dict get $rep_info name] [dict get $rep_info ver] $pkg_inst_dir $tmp_build_dir
+	}
+	if {$run_check && $do_check} {
+		puts "Running check{}"
+		check $short_name $pkg_inst_dir $tmp_build_dir
+	}
+	if {$run_install} {
+		puts "Running install{}"
+		install $short_name $pkg_inst_dir $tmp_build_dir
+	}
+	set inst_files [glob -nocomplain -directory $inst_dir *]]
+	if {[llength $inst_files] == 0} {
+		error "This recipe didn't install any files!"
+	}
 
-    # Dump the hash to see if the recipe's changed later.
-    set hash_file [file join $pkg_inst_dir recipe-hash.txt]
-    set hash_file [open $hash_file w]
-    puts $hash_file $src_hash
-    close $hash_file
+	# Dump the hash to see if the recipe's changed later.
+	set hash_file [file join $pkg_inst_dir recipe-hash.txt]
+	set hash_file [open $hash_file w]
+	puts $hash_file $src_hash
+	close $hash_file
 
-    array set ::env $old_env
+	array set ::env $old_env
 
 	# TODO: Delete the install dir if the install fails
 	file delete -force -- $tmp_build_dir
 
 	puts "\nInstalled $short_name"
-    lappend built_pkgs $short_name
-    return $built_pkgs
+	lappend built_pkgs $short_name
+	return $built_pkgs
 }
 
 if {[llength $build_reps] > 0} {
